Update test runner to log test output
It's still a little rough around the edges, but this patch updates the
test runner used by the Fortran tests to capture output in a log file
and dump the output when a test fails.  This is extremely useful for
testing that's run in our continuous integration environment where we
can't run interactive testing or access log files after a round of
testing completes.

Signed-off-by: James Dinan <james.dinan@intel.com>