more fixes for Cray and Sandia OpenSHMEM
Turns out the logic was wrong in the tests in this commit.

The tests were only working because the default for some
OpenSHMEM's is to do strict error checking, or that the
underlying transport layer melts down when the target of
a put is not in the symmetric heap or data segment.

Fix this so now the tests fail when compiled against
Cray and OpenSHMEM.

Oh, and by the way, none of the fortran tests in this repo
can be compiled with the Cray compile wrappers/cray-shmem
module unless they are modified.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>