Update ARGV check
The current ARGV check 'defined(@array)' is deprecated in Perl; replace
this with '@ARGV'.  When the array is referenced in a scalar context
(e.g. if(@ARGV)), the expression will evaluate to the length of the
array, which is perhaps (I mean, who reallys knows, it's Perl) a safer
way to check for args anyway.

Signed-off-by: James Dinan <james.dinan@intel.com>